+++ src/containers/App/App.jsx
@@ -26,6 +26,7 @@ export const App = React.createClass({
 		loginData: PropTypes.object,
 		path: PropTypes.string,
 		slug: PropTypes.string,
+		query: PropTypes.object,
 		children: PropTypes.object,
 		dispatch: PropTypes.func
 	},
@@ -82,6 +83,7 @@ export const App = React.createClass({
 		const notFound = safeGetInToJS(this.props.appData, ['notFound']) || !isLoggedIn && this.props.path.substring(this.props.path.length - 9, this.props.path.length) === '/settings' || false;
 		const messages = safeGetInToJS(this.props.appData, ['languageObject']) || {}; // Messages includes all of the strings used on the site. Language support is implemented by sending a different messages object.
 		const hideFooter = notFound || this.props.path.substring(this.props.path.length - 6, this.props.path.length) === '/draft' || this.props.path.substring(this.props.path.length - 6, this.props.path.length) === '/login' || this.props.path.substring(this.props.path.length - 7, this.props.path.length) === '/signup' || this.props.path.substring(0, 7) === '/verify'; // We want to hide the footer if we are in the editor or login. All other views show the footer.
+		const isEmbed = this.props.query && this.props.query.embed;
 		const metaData = { // Metadata that will be used by Helmet to populate the <head> tag
 			meta: [
 				{property: 'og:site_name', content: 'PubPub'},
@@ -98,7 +100,12 @@ export const App = React.createClass({
 
 					<Helmet {...metaData} />
 					<AppLoadingBar color={'#BBBDC0'} show={this.props.appData.get('loading')} />
-					<AppHeader loginData={this.props.loginData} path={this.props.path} createDocument={this.createDocument} logoutHandler={this.logoutHandler} goToURL={this.goToURL}/>
+
+					{!isEmbed &&
+						<AppHeader loginData={this.props.loginData} path={this.props.path} createDocument={this.createDocument} logoutHandler={this.logoutHandler} goToURL={this.goToURL}/>
+					}
+					
+
 					<AppVerified isVerified={!notVerified} handleResendEmail={this.handleResendEmail}/>
 					<AppMessage/>
 					{notFound && <NotFound />}
@@ -119,6 +126,7 @@ export default connect( state => {
 		loginData: state.login,
 		mediaData: state.media,
 		path: state.router.location.pathname,
+		query: state.router.location.query,
 		slug: state.router.params.slug,
 	};
 })( App );